moving functions into components
